REFACTOR: se refactorizó la aplicación organizando los modelos y servicios
